Update to 1.17 and remove support for older versions

+++ src/main/java/com/dutchjelly/craftenhance/gui/guis/CustomCraftingTable.java
@@ -6,20 +6,17 @@ import com.dutchjelly.craftenhance.gui.GuiManager;
 import com.dutchjelly.craftenhance.gui.templates.GuiTemplate;
 import com.dutchjelly.craftenhance.gui.util.GuiUtil;
 import com.dutchjelly.craftenhance.messaging.Debug;
-import org.apache.commons.lang.math.IntRange;
 import org.bukkit.Bukkit;
-import org.bukkit.Location;
-import org.bukkit.Material;
-import org.bukkit.entity.HumanEntity;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.ClickType;
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.event.inventory.InventoryDragEvent;
-import org.bukkit.event.inventory.InventoryType;
-import org.bukkit.inventory.*;
+import org.bukkit.inventory.Inventory;
+import org.bukkit.inventory.ItemStack;
 
-import java.lang.reflect.Array;
-import java.util.*;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 public class CustomCraftingTable extends GUIElement {
@@ -94,7 +91,7 @@ public class CustomCraftingTable extends GUIElement {
                     false, null
             );
             int space = destination.keySet().stream().filter(x -> x != null && x != -1)
-                    .mapToInt(x -> destination.get(x).intValue()).sum();
+                    .mapToInt(x -> destination.get(x)).sum();
             canCraft = Math.min(findFittingRecipesAmount(), space);
 
             if(canCraft == 0){
